+++ frontend/src/components/Round.js
@@ -1,117 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import { View, Text, TouchableOpacity, Image, StyleSheet } from 'react-native'
 import scheduleData from '../data/schedule.json'
-
-const Timeline = ({ roundData }) => {
-  const parseTime = (stopTime) => {
-    const [hours, minutes] = stopTime.split(':')
-    return {
-      hours: parseInt(hours),
-      minutes: parseInt(minutes),
-    }
-  }
-
-  const isPast = (stopTime) => {
-    const currentTime = new Date()
-    const { hours, minutes } = parseTime(stopTime)
-    const stopDate = new Date(
-      currentTime.getFullYear(),
-      currentTime.getMonth(),
-      currentTime.getDate(),
-      hours,
-      minutes
-    )
-    return currentTime > stopDate
-  }
-
-  const isCurrent = (stopTime, nextStopTime) => {
-    const currentTime = new Date()
-    const { hours, minutes } = parseTime(stopTime)
-    const stopDate = new Date(
-      currentTime.getFullYear(),
-      currentTime.getMonth(),
-      currentTime.getDate(),
-      hours,
-      minutes
-    )
-    if (!nextStopTime) {
-      // If there's no next stop time, revert to the original logic
-      const nextMinutes = new Date(stopDate.getTime() + 60000) // Adds 1 minute to stop time
-      return currentTime >= stopDate && currentTime < nextMinutes
-    } else {
-      const { hours: nextHours, minutes: nextMinutes } = parseTime(nextStopTime)
-      const nextStopDate = new Date(
-        currentTime.getFullYear(),
-        currentTime.getMonth(),
-        currentTime.getDate(),
-        nextHours,
-        nextMinutes
-      )
-      // Now checks if current time is before the next stop's time
-      return currentTime >= stopDate && currentTime < nextStopDate
-    }
-  }
-
-  return (
-    <View style={styles.timelineContainer}>
-      {roundData.map((stop, index) => (
-        <View key={index} style={styles.stopContainer}>
-          <View style={styles.lineAndCircle}>
-            {index !== 0 && (
-              <View
-                style={[
-                  styles.verticalLine,
-                  isCurrent(stop.time, roundData[index + 1]?.time)
-                    ? styles.current
-                    : isPast(stop.time)
-                    ? styles.past
-                    : styles.future,
-                ]}
-              />
-            )}
-            <View
-              style={[
-                styles.circle,
-                isCurrent(stop.time, roundData[index + 1]?.time)
-                  ? styles.current
-                  : isPast(stop.time)
-                  ? styles.past
-                  : styles.future,
-              ]}
-            />
-            {index !== roundData.length - 1 && (
-              <View
-                style={[
-                  styles.verticalLine,
-                  isCurrent(stop.time, roundData[index + 1]?.time)
-                    ? styles.current
-                    : isPast(stop.time)
-                    ? styles.past
-                    : styles.future,
-                ]}
-              />
-            )}
-          </View>
-          <View style={styles.stopDetails}>
-            <Text
-              style={[
-                styles.stopTime,
-                isCurrent(stop.time, roundData[index + 1]?.time)
-                  ? styles.currentTime
-                  : isPast(stop.time)
-                  ? styles.pastText
-                  : styles.futureText,
-              ]}
-            >
-              {stop.time}
-            </Text>
-            <Text style={styles.stopName}>{stop.name}</Text>
-          </View>
-        </View>
-      ))}
-    </View>
-  )
-}
+import Timeline from './Timeline'
 
 const Round = ({ isOperationDay }) => {
   const [currentTime, setCurrentTime] = useState(new Date())
@@ -130,9 +20,6 @@ const Round = ({ isOperationDay }) => {
     if (index === -1) index = 9
     setCurrentIndex(index)
   }, [currentTime])
-
-  const currentRound = scheduleData.schedule[selectedIndex]
-  const { round, tables } = currentRound
 
   const getCurrentIndex = () => {
     const currentHour = currentTime.getHours()
@@ -188,6 +75,9 @@ const Round = ({ isOperationDay }) => {
       </View>
     )
   }
+
+  const currentRound = scheduleData.schedule[selectedIndex]
+  const { round, tables } = currentRound
 
   return (
     <View style={styles.container}>
@@ -267,60 +157,10 @@ const styles = StyleSheet.create({
     marginBottom: 10,
     color: '#333333', // Default color
   },
-  timelineContainer: {
-    flexDirection: 'column',
-    alignItems: 'flex-start',
-    justifyContent: 'flex-start',
-    marginTop: 20,
-    marginLeft: 20,
+  image: {
+    width: 200,
+    height: 200,
   },
-  stopContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-  lineAndCircle: {
-    alignItems: 'center',
-    marginRight: 10,
-  },
-  verticalLine: {
-    width: 2,
-    backgroundColor: '#000',
-    flex: 1, // Take up available space
-  },
-  circle: {
-    width: 10,
-    height: 10,
-    borderRadius: 5,
-    backgroundColor: '#000',
-    marginBottom: 2, // Spacing between circle and line
-  },
-  stopDetails: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  stopTime: {
-    marginRight: 10,
-    fontWeight: 'bold',
-  },
-  stopName: {
-    fontSize: 16,
-  },
-  past: {
-    backgroundColor: '#808080', // Gray for past times
-  },
-  future: {
-    backgroundColor: '#39FF14', // Fluorescent green for future times
-  },
-  pastText: {
-    color: '#808080', // Gray text for past times
-    fontWeight: 'bold',
-  },
-  futureText: {
-    color: '#39FF14', // Fluorescent green text for future times
-    fontWeight: 'bold',
-  },
-  // Include other existing styles here
 })
 
 export default Round
